Update filehandles inventory function
Changed inventory function to use the Checkmk host_extra_conf() function instead of manually parsing the ruleset.
Accessing a rule dictionary object with array indexes (e.g. rule[1:3]) leads to an "unhashable type" error in Python 2.7 and below, as the rules are non-ordered dictionaries. Dictionaries are only ordered by default starting with Python 3.6, which is not the Python version shipped with Checkmk 1.6.